added keys API to manage the import and export of encrypted keys, This uses the YAJL subproject from the  https://github.com/lloyd/yajl.git library, and thus needed some tweeking of the project file to build YAJL unchanged.
Keys can now be serialized (encoded to a passphrase) and deserialized and then decrypted.

the new APIs' are C4Key_NewSymmetric, C4Key_NewTBC, C4Key_Free which manipulate the C4KeyContext objects
and the C4Key_SerializeToPassPhrase which produces JSON that can be read back in with C4Key_Deserialize
and then checked with C4Key_VerifyPassPhrase or decrypted back to a  C4KeyContext object using C4Key_DecryptFromPassPhrase.