added three new APIs to S4 to properly handle the encoding of keys with passwords.
one of the issues we should address is that there are databases out there of common passphrase
encoded with the various P2K algorithms .  we should strengthen our passphrase before we actually
send them over the net and infant when we use them to encode out database..

A simple way to do this is to hash them with a salt.

I have added the HASH_NormalizePassPhrase Api to simplify this.

err = HASH_NormalizePassPhrase(passphrase,passphraseLen,
							   (uint8_t*)saltString, strlen(saltString),
							   &passCode, &passCodeLen); CKERR;

given a  passphrase and a salt we will do  the following

 passcode = SHA_3_512 ( passphraseLen(16) | passphraseBytes | saltLength(16) } saltBytes )

which returns is a 64 bytes block resulting from the above hash.

for cloud files, we could consider the file cloud name or maybe the owner bucketID as the salt.
as long as it doesn’t change  for encoding and decoding.

———————————

given a key of arbitrary length and a passcode typically strengthen from above.  we then can
run an encrypt function

where we can specify the key and it’s length and the cipher wish to encode with along
with a passcode and a passphrase encoding algorithm

valid ciphers are AES-128, AES-256, Twofish-256

value passPhraseAlgorithm functions are: PBKDF2, Argon2d, Argon2i, Argon2id

 err = S4Key_EncryptKeyToPassPhrase( key, sizeof(key), cipherAlgorithm,
						   passCode, passCodeLen, passPhraseAlgorithm,
						   &eskData, &eskDataLen);

given these inputs we get back a JSON block something like

{
    "version": 1,
    "encoding": "p2k",
    "keySuite": "Twofish-256",
    "p2k-params": "$Argon2i$m=65536,t=2,p=1,k=32$R8XS+TpeaWU=",
    "mac": "LwV7sFBjrlE=",
    "esk": "6BTAgOheYJZmOalbK2eDpoY5RQMqZ6APmf5KAeIEyiQ=",
    "iv": "N++5eulCre9WpDvoItRjx76n18skjWUT9SQdAgj2k3w=",
    "encrypted": "aS/NdU/ie32tbWkhTzh8UFTXWlASrbM...7zFSHrswUvWNkVoD+s2Vbi6HBwgZ0y4="
}

there is a corresponding decode function that takes this block and hands you back a the original key
We check for passcode validity using a mac function, so if you type in a bad passcode you will
error out  from the function

err = S4Key_DecryptKeyFromPassPhrase(eskData, eskDataLen,
						 passCode, passCodeLen,
						 &key1, & key1Len); CKERR;